backend/requirements.txt
apscheduler
cryptography
fastapi
gradio==3.32.0
httpx
minio
openai
opencc-python-reimplemented
Pillow
pydantic[email]
pymysql
python-dotenv
python-multipart
PyJWT
redis
stability-sdk
SQLAlchemy
uvicorn[standard]